requirements.txt
ffmpeg-python>=0.2.0
scenedetect[opencv]>=0.6.1